First approach

+++ shutter/scripts/configure_shuttermint.sh
@@ -3,16 +3,23 @@
 # shellcheck disable=SC1091
 . "${ASSETS_DIR}/variables.env"
 
+if [ ! -f "$SHUTTER_CHAIN_CONFIG_FILE" ]; then
+    echo "[ERROR | configure] Missing chain configuration file (${SHUTTER_CHAIN_CONFIG_FILE})"
+    exit 1
+fi
+
 ASSETS_GENESIS_FILE="/assets/genesis.json"
 CHAIN_GENESIS_FILE="${SHUTTER_CHAIN_DIR}/config/genesis.json"
 
 rm "$CHAIN_GENESIS_FILE"
 ln -s "$ASSETS_GENESIS_FILE" "$CHAIN_GENESIS_FILE"
 
-SHUTTERMINT_MONIKER=${KEYPER_NAME:-$(openssl rand -hex 8)}
+# KEYPER_NAME=${KEYPER_NAME:-$(openssl rand -hex 8)}
+
+# TODO: Call the go binary
 
 sed -i "/^seeds =/c\seeds = \"${_ASSETS_SHUTTERMINT_SEED_NODES}\"" "$SHUTTER_CHAIN_CONFIG_FILE"
-sed -i "/^moniker =/c\moniker = \"${SHUTTERMINT_MONIKER}\"" "$SHUTTER_CHAIN_CONFIG_FILE"
+sed -i "/^moniker =/c\moniker = \"${KEYPER_NAME}\"" "$SHUTTER_CHAIN_CONFIG_FILE"
 sed -i "/^genesis_file =/c\genesis_file = \"${ASSETS_GENESIS_FILE}\"" "$SHUTTER_CHAIN_CONFIG_FILE"
 sed -i "/^external_address =/c\external_address = \"${_DAPPNODE_GLOBAL_PUBLIC_IP}:${CHAIN_PORT}\"" "$SHUTTER_CHAIN_CONFIG_FILE"
 sed -i "/^addr_book_strict =/c\addr_book_strict = true" "$SHUTTER_CHAIN_CONFIG_FILE"
